# swagger.json
{"openapi":"3.0.0","paths":{"/pokemon":{"post":{"operationId":"PokemonController_create","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreatePokemonDto"}}}},"responses":{"201":{"description":""}},"tags":["Create one pokemon"]},"get":{"operationId":"PokemonController_findAll","parameters":[{"name":"take","required":false,"in":"query","schema":{"type":"number"}},{"name":"skip","required":false,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"tags":["Get all pokemons"]}},"/pokemon/{name}":{"get":{"operationId":"PokemonController_findOne","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Get one pokemon"]},"patch":{"operationId":"PokemonController_update","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdatePokemonDto"}}}},"responses":{"200":{"description":""}},"tags":["Update one pokemon"]},"delete":{"operationId":"PokemonController_remove","parameters":[{"name":"name","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"tags":["Remove one pokemmon"]}}},"info":{"title":"Pokemon","description":"Small API Restful application exposing CRUD operations for Pokemons management","version":"1.0","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreatePokemonDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"type1":{"type":"string"},"type2":{"type":"string"},"Total":{"type":"number"},"HP":{"type":"number"},"Attack":{"type":"number"},"Defense":{"type":"number"},"SpAtk":{"type":"number"},"SpDef":{"type":"number"},"Speed":{"type":"number"},"Generation":{"type":"number"}},"required":["id","name","type1","type2","Total","HP","Attack","Defense","SpAtk","SpDef","Speed","Generation"]},"UpdatePokemonDto":{"type":"object","properties":{"id":{"type":"number"},"name":{"type":"string"},"type1":{"type":"string"},"type2":{"type":"string"},"Total":{"type":"number"},"HP":{"type":"number"},"Attack":{"type":"number"},"Defense":{"type":"number"},"SpAtk":{"type":"number"},"SpDef":{"type":"number"},"Speed":{"type":"number"},"Generation":{"type":"number"}}}}}}